Handle the case when there are no knative crds

+++ hack/lib/serverless.bash
@@ -358,7 +358,9 @@ function teardown_serverless {
   done
   oc delete namespace openshift-serverless --ignore-not-found=true
 
-  oc get crd -oname | grep 'knative.dev' | xargs oc delete --timeout=60s
+  if [[ ! $(oc get crd -oname | grep -c 'knative.dev') -eq 0 ]]; then
+    oc get crd -oname | grep 'knative.dev' | xargs oc delete --timeout=60s
+  fi
 
   logger.success 'Serverless has been uninstalled.'
 }
